requirements.txt
django
djangorestframework
slackclient==1.3.1
slackeventsapi
slack-sdk
asgiref==3.7.2
sqlparse==0.4.4
typing_extensions==4.9.0
langchain
openai
langchain-openai
langchain-community
psycopg
psycopg2
python-dotenv
pandas
SQLAlchemy
django-crontab